requirements.txt
asgiref==3.8.1
Django==5.1.7
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
redis==6.0.0
sqlparse==0.5.3
tzdata==2025.1
